Fixed sockgetremotename.
git-svn-id: svn+ssh://svn.dolda2000.com/srv/svn/repos/src/doldaconnect@730 959494ce-11ee-0310-bf91-de5d638817bd

+++ daemon/ui.c
@@ -144,7 +144,7 @@ struct uidata
     size_t inbufsize, indata;
     /* Wordset storage */
     wchar_t **argv;
-    int argc, args;
+    size_t argc, args;
     /* WCS conversation stuff */
     wchar_t *cb; /* Conversation buffer */
     size_t cbsize, cbdata;
